Move spectral error plt func
