+++ fooof/tests/plts/test_spectra.py
@@ -48,11 +48,3 @@ def test_plot_spectra_shading(tfg, skip_if_no_mpl):
     plot_spectra_shading(tfg.freqs, [tfg.power_spectra[0, :], tfg.power_spectra[1, :]],
                          shades=[8, 12], add_center=True, log_freqs=True, log_powers=True,
                          labels=['A', 'B'])
-
-@plot_test
-def test_plot_spectral_error(skip_if_no_mpl):
-
-    fs = np.arange(3, 41, 1)
-    errs = np.ones(len(fs))
-
-    plot_spectral_error(fs, errs)
